Before Fixing GAP between Conveyors
Fixed Errors:
Level Retry, Next Level Buttons delay Active LOGIC.   ( On Gameplay & Tutorial Scene )
Play a Random Level after Completing all Levels.
Fixed Conveyor VALID placing color to ORIGINAL instead of GREEN.

+++ Assets/Scripts/Conveyor/ConveyorExitController.cs
@@ -5,7 +5,7 @@ public class ConveyorExitController : MonoBehaviour
 
     public Machine connectedMachine;
     public Conveyor _conveyor;
-    public int OutForce = 2;
+    public float OutForce = 2;
 
 
     void Awake()
